Core data fixes

+++ Sources/EmbraceUploadInternal/Cache/EmbraceUploadCache.swift
@@ -90,7 +90,7 @@ class EmbraceUploadCache {
 
         // update if it already exists
         if let record = fetchUploadData(id: id, type: type) {
-            coreData.context.perform { [weak self] in
+            coreData.context.performAndWait { [weak self] in
                 record.data = data
                 self?.coreData.save()
             }
@@ -105,19 +105,22 @@ class EmbraceUploadCache {
         var result = true
 
         coreData.context.performAndWait {
-            let record = UploadDataRecord.create(
+            if let record = UploadDataRecord.create(
                 context: coreData.context,
                 id: id,
                 type: type.rawValue,
                 data: data,
                 attemptCount: 0,
                 date: Date()
-            )
+            ) {
 
-            do {
-                try coreData.context.save()
-            } catch {
-                coreData.context.delete(record)
+                do {
+                    try coreData.context.save()
+                } catch {
+                    coreData.context.delete(record)
+                    result = false
+                }
+            } else {
                 result = false
             }
         }
@@ -132,7 +135,7 @@ class EmbraceUploadCache {
             return
         }
 
-        coreData.context.perform { [weak self] in
+        coreData.context.performAndWait { [weak self] in
             guard let strongSelf = self else {
                 return
             }
@@ -183,7 +186,7 @@ class EmbraceUploadCache {
             return
         }
 
-        coreData.context.perform { [weak self] in
+        coreData.context.performAndWait { [weak self] in
             uploadData.attemptCount = attemptCount
             self?.coreData.save()
         }
